+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.2
+
+- Highlight bracket pairs
+- Added config param: `"clojureWarrior.matchedBracketStyle": {"backgroundColor": "#E0E0E0"}`
+
 # 0.1.1
 
 Added configuration parameters: